+++ src/sourcegraph/stream-search/index.ts
@@ -83,10 +83,18 @@ export async function performSearch(
       handlers.onResults(
         event.data.map((match): SearchResult => {
           const url = `${src.instance}${getMatchUrl(match)}`;
-          if (match.type === "commit") {
-            // Commit stuff comes already markdown-formatted?? so strip formatting
-            match.label = stripMarkdown.processSync(match.label)?.value.toString().split(`› `).pop() || "";
-            match.detail = stripMarkdown.processSync(match.detail)?.value.toString();
+          switch (match.type) {
+            case "commit":
+              // Commit stuff comes already markdown-formatted?? so strip formatting
+              match.label = stripMarkdown.processSync(match.label)?.value.toString().split(`› `).pop() || "";
+              match.detail = stripMarkdown.processSync(match.detail)?.value.toString();
+              break;
+            case "content":
+              // Line number appears 0-indexed, for ease of use increment it so links
+              // aren't off by 1.
+              match.lineMatches.forEach((l) => {
+                l.lineNumber += 1;
+              });
           }
           return { url, match };
         })